+++ Release Application/RoSA_Monitor_autogen/EWIEGA46WW/moc_rosawindow.cpp
@@ -22,8 +22,8 @@ QT_BEGIN_MOC_NAMESPACE
 QT_WARNING_PUSH
 QT_WARNING_DISABLE_DEPRECATED
 struct qt_meta_stringdata_RosaWindow_t {
-    QByteArrayData data[18];
-    char stringdata0[391];
+    QByteArrayData data[20];
+    char stringdata0[441];
 };
 #define QT_MOC_LITERAL(idx, ofs, len) \
     Q_STATIC_BYTE_ARRAY_DATA_HEADER_INITIALIZER_WITH_OFFSET(len, \
@@ -49,7 +49,9 @@ QT_MOC_LITERAL(13, 269, 26), // "on_selectMapButton_clicked"
 QT_MOC_LITERAL(14, 296, 21), // "on_SlamButton_clicked"
 QT_MOC_LITERAL(15, 318, 27), // "on_navigationButton_clicked"
 QT_MOC_LITERAL(16, 346, 22), // "on_esp32Button_clicked"
-QT_MOC_LITERAL(17, 369, 21) // "on_rosaButton_clicked"
+QT_MOC_LITERAL(17, 369, 21), // "on_rosaButton_clicked"
+QT_MOC_LITERAL(18, 391, 25), // "on_nodeListButton_clicked"
+QT_MOC_LITERAL(19, 417, 23) // "on_cameraButton_clicked"
 
     },
     "RosaWindow\0windowClosed\0\0"
@@ -63,7 +65,9 @@ QT_MOC_LITERAL(17, 369, 21) // "on_rosaButton_clicked"
     "on_TeleopButton_clicked\0"
     "on_selectMapButton_clicked\0"
     "on_SlamButton_clicked\0on_navigationButton_clicked\0"
-    "on_esp32Button_clicked\0on_rosaButton_clicked"
+    "on_esp32Button_clicked\0on_rosaButton_clicked\0"
+    "on_nodeListButton_clicked\0"
+    "on_cameraButton_clicked"
 };
 #undef QT_MOC_LITERAL
 
@@ -73,7 +77,7 @@ static const uint qt_meta_data_RosaWindow[] = {
        8,       // revision
        0,       // classname
        0,    0, // classinfo
-      16,   14, // methods
+      18,   14, // methods
        0,    0, // properties
        0,    0, // enums/sets
        0,    0, // constructors
@@ -81,29 +85,33 @@ static const uint qt_meta_data_RosaWindow[] = {
        1,       // signalCount
 
  // signals: name, argc, parameters, tag, flags
-       1,    0,   94,    2, 0x06 /* Public */,
+       1,    0,  104,    2, 0x06 /* Public */,
 
  // slots: name, argc, parameters, tag, flags
-       3,    0,   95,    2, 0x08 /* Private */,
-       4,    0,   96,    2, 0x08 /* Private */,
-       5,    0,   97,    2, 0x08 /* Private */,
-       6,    0,   98,    2, 0x08 /* Private */,
-       7,    0,   99,    2, 0x08 /* Private */,
-       8,    0,  100,    2, 0x08 /* Private */,
-       9,    0,  101,    2, 0x08 /* Private */,
-      10,    0,  102,    2, 0x08 /* Private */,
-      11,    0,  103,    2, 0x08 /* Private */,
-      12,    0,  104,    2, 0x08 /* Private */,
-      13,    0,  105,    2, 0x08 /* Private */,
-      14,    0,  106,    2, 0x08 /* Private */,
-      15,    0,  107,    2, 0x08 /* Private */,
-      16,    0,  108,    2, 0x08 /* Private */,
-      17,    0,  109,    2, 0x08 /* Private */,
+       3,    0,  105,    2, 0x08 /* Private */,
+       4,    0,  106,    2, 0x08 /* Private */,
+       5,    0,  107,    2, 0x08 /* Private */,
+       6,    0,  108,    2, 0x08 /* Private */,
+       7,    0,  109,    2, 0x08 /* Private */,
+       8,    0,  110,    2, 0x08 /* Private */,
+       9,    0,  111,    2, 0x08 /* Private */,
+      10,    0,  112,    2, 0x08 /* Private */,
+      11,    0,  113,    2, 0x08 /* Private */,
+      12,    0,  114,    2, 0x08 /* Private */,
+      13,    0,  115,    2, 0x08 /* Private */,
+      14,    0,  116,    2, 0x08 /* Private */,
+      15,    0,  117,    2, 0x08 /* Private */,
+      16,    0,  118,    2, 0x08 /* Private */,
+      17,    0,  119,    2, 0x08 /* Private */,
+      18,    0,  120,    2, 0x08 /* Private */,
+      19,    0,  121,    2, 0x08 /* Private */,
 
  // signals: parameters
     QMetaType::Void,
 
  // slots: parameters
+    QMetaType::Void,
+    QMetaType::Void,
     QMetaType::Void,
     QMetaType::Void,
     QMetaType::Void,
@@ -145,6 +153,8 @@ void RosaWindow::qt_static_metacall(QObject *_o, QMetaObject::Call _c, int _id, 
         case 13: _t->on_navigationButton_clicked(); break;
         case 14: _t->on_esp32Button_clicked(); break;
         case 15: _t->on_rosaButton_clicked(); break;
+        case 16: _t->on_nodeListButton_clicked(); break;
+        case 17: _t->on_cameraButton_clicked(); break;
         default: ;
         }
     } else if (_c == QMetaObject::IndexOfMethod) {
@@ -189,13 +199,13 @@ int RosaWindow::qt_metacall(QMetaObject::Call _c, int _id, void **_a)
     if (_id < 0)
         return _id;
     if (_c == QMetaObject::InvokeMetaMethod) {
-        if (_id < 16)
+        if (_id < 18)
             qt_static_metacall(this, _c, _id, _a);
-        _id -= 16;
+        _id -= 18;
     } else if (_c == QMetaObject::RegisterMethodArgumentMetaType) {
-        if (_id < 16)
+        if (_id < 18)
             *reinterpret_cast<int*>(_a[0]) = -1;
-        _id -= 16;
+        _id -= 18;
     }
     return _id;
 }