+++ RoSA_Monitor/rosawindow.cpp
@@ -112,7 +112,7 @@ void RosaWindow::on_rosaButton_clicked()
                }
                if(checkBox->objectName() == "cameraCheckbox")
                {
-                    //TO DO
+                    on_cameraButton_clicked();
                }
            }
        }
@@ -139,7 +139,7 @@ void RosaWindow::on_URDFButton_clicked()
 {
     if(ButtonPressed(ui->URDFButton, LauncherManager::URDF, "URDF"))
     {
-        qDebug()<<"Firmware correctly executed";
+        qDebug()<<"URDF executed";
     }
     else
     {
@@ -153,7 +153,7 @@ void RosaWindow::on_lidarButton_clicked()
 {
     if(ButtonPressed(ui->lidarButton, LauncherManager::LIDAR, "Hokuyo LiDAR"))
     {
-        qDebug()<<"Firmware correctly executed";
+        qDebug()<<"LiDAR correctly executed";
     }
     else
     {
@@ -163,6 +163,19 @@ void RosaWindow::on_lidarButton_clicked()
     }
 }
 
+void RosaWindow::on_cameraButton_clicked()
+{
+    if(ButtonPressed(ui->cameraButton, LauncherManager::CAMERA, "RGBD Camera"))
+    {
+        qDebug()<<"Camera correctly executed";
+    }
+    else
+    {
+        qDebug()<<"Error executing RGBD Camera";
+        ui->cameraButton->setChecked(false);
+
+    }
+}
 
 void RosaWindow::on_SlamButton_clicked()
 {
@@ -505,6 +518,72 @@ void RosaWindow::on_mainMenuButton_2_clicked()
     on_mainMenuButton_clicked();
 }
 
+void RosaWindow::on_nodeListButton_clicked()
+{
+    //Ask user for a workspace if not selected yet
+    if(manager.GetWorkspacePath()==nullptr)
+    {
+        QMessageBox::information(this, tr("ROSA Info"), tr("You have to select a ROS2 workspace first"));
+        QString workspace_name = "";
+
+        workspace_name= QFileDialog::getExistingDirectory(this, "Select your workspace", QDir::homePath(), QFileDialog::ShowDirsOnly);
+        if(workspace_name == "")
+        {
+            ui->statusbar->showMessage("No workspace selected", 5000);
+            return;
+        }
+        SetWorkspace(workspace_name);
+
+        //Verify if /maps exists
+        QString mapsDirPath = manager.GetWorkspacePath()->filePath("maps");
+        bool mapsExist = manager.GetWorkspacePath()->exists("maps");
+
+        //If not, create /maps in workspace
+        if (!mapsExist)
+        {
+            manager.GetWorkspacePath()->mkpath(mapsDirPath);
+            ui->statusbar->showMessage("Maps folder created", 5000);
+        }
+    }
+    if(manager.GetLauncher(LauncherManager::NODE_LIST)->GetActive())
+    {
+        manager.StopLauncher(LauncherManager::NODE_LIST);
+    }
+    //Create new process
+    if(!manager.CreateLauncher(LauncherManager::NODE_LIST, new QProcess(this)))
+    {
+        ui->statusbar->showMessage("Launch failed", 5000);
+        return;
+    }
+    else
+    {
+            //if first time launched create new tab
+        if(shellWindows[LauncherManager::NODE_LIST] == nullptr)
+        {
+            //Create shell window and connect shell output of process launcher to the read output slot
+            shellWindows[LauncherManager::NODE_LIST] = new ShellOutputWindow(this, manager.GetLauncher(LauncherManager::NODE_LIST)->GetLauncherProcess());
+            //Add new tab
+            ui->tabRoSAWidget->addTab(shellWindows[LauncherManager::NODE_LIST], "ROS2 node list");
+            tab_Index[LauncherManager::NODE_LIST] = numOfTabs++;
+
+            ui->nodeListButton->setText("Update ROS2 node list");
+        }
+        //if not first time clear tab and associates the new process
+        else
+        {
+            shellWindows[LauncherManager::NODE_LIST]->ClearText();
+            shellWindows[LauncherManager::NODE_LIST]->setLauncherProcess(manager.GetLauncher(LauncherManager::NODE_LIST)->GetLauncherProcess());
+        }
+        connect(manager.GetLauncher(LauncherManager::NODE_LIST)->GetLauncherProcess(), &QProcess::readyReadStandardOutput, shellWindows[LauncherManager::NODE_LIST], &ShellOutputWindow::readProcessOutput);
+
+        //Launch process
+        manager.Launch(LauncherManager::NODE_LIST);
+
+        ui->statusbar->showMessage("Launching node list", 5000);
+    }
+}
+
+
 
 
 /****************************************************
@@ -657,6 +736,3 @@ void RosaWindow::RemoveTab(LauncherManager::LauncherType type)
 
     tab_Index[type] = 0;
 }
-
-
-